+++ src/core/components/dialogs/dialogAlert/dialogAlertRoot/dialogAlertRoot.stories.tsx
@@ -22,9 +22,6 @@ type Story = StoryObj<typeof DialogAlert.Root>;
  * Default usage of the `DialogAlert.Root` component
  */
 export const Default: Story = {
-    args: {
-        size: 'md',
-    },
     render: (props) => {
         const [open, setOpen] = useState(false);